+++ src/main/java/com/petspace/dev/service/ReviewService.java
@@ -10,8 +10,7 @@ import com.petspace.dev.dto.review.ReviewCreateResponseDto;
 import com.petspace.dev.repository.ReviewImageRepository;
 import com.petspace.dev.repository.ReservationRepository;
 import com.petspace.dev.repository.UserRepository;
-import com.petspace.dev.util.exception.handler.AwsException;
-import com.petspace.dev.util.exception.handler.ReviewException;
+import com.petspace.dev.util.exception.ReviewException;
 import com.petspace.dev.util.s3.AwsS3Uploader;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
