+++ src/main/java/com/petspace/dev/util/exception/AwsException.java
@@ -1,14 +1,14 @@
-package com.petspace.dev.util.exception.handler;
+package com.petspace.dev.util.exception;
 
 import com.petspace.dev.util.BaseResponseStatus;
 import lombok.Getter;
 
 @Getter
-public class ReviewException extends RuntimeException{
+public class AwsException extends RuntimeException{
 
     private final BaseResponseStatus status;
 
-    public ReviewException(BaseResponseStatus status) {
+    public AwsException(BaseResponseStatus status) {
         this.status = status;
     }
 }
