+++ src/main/java/com/petspace/dev/util/s3/AwsS3Uploader.java
@@ -4,14 +4,13 @@ import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.petspace.dev.util.BaseResponse;
-import com.petspace.dev.util.exception.handler.ReviewException;
+import com.petspace.dev.util.exception.ReviewException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
-import com.petspace.dev.util.exception.handler.AwsException;
+import com.petspace.dev.util.exception.AwsException;
 
 import java.io.File;
 import java.io.FileOutputStream;
